OpenAIChatCompletionAgent removed to simplify the design
