+++ dotnet/src/Experimental/Agents/ChatCompletion/ChatCompletionAgent.cs
@@ -11,18 +11,15 @@ namespace Microsoft.SemanticKernel.Experimental.Agents;
 /// <summary>
 /// Represent an agent that is built around the SK ChatCompletion API and leverages the API's capabilities.
 /// </summary>
-public abstract class ChatCompletionAgent : KernelAgent
+public sealed class ChatCompletionAgent : KernelAgent
 {
-    private readonly string _instructions;
-    private readonly PromptExecutionSettings? _promptExecutionSettings;
-
     /// <summary>
     /// Initializes a new instance of the <see cref="ChatCompletionAgent"/> class.
     /// </summary>
     /// <param name="kernel">The <see cref="Kernel"/> containing services, plugins, and other state for use by the agent.</param>
     /// <param name="instructions">The instructions for the agent.</param>
     /// <param name="executionSettings">The optional execution settings for the agent. If not provided, default settings will be used.</param>
-    protected ChatCompletionAgent(Kernel kernel, string instructions, PromptExecutionSettings? executionSettings = null) : base(kernel)
+    public ChatCompletionAgent(Kernel kernel, string instructions, PromptExecutionSettings? executionSettings = null) : base(kernel)
     {
         Verify.NotNullOrWhiteSpace(instructions, nameof(instructions));
         this._instructions = instructions;
@@ -36,17 +33,30 @@ public abstract class ChatCompletionAgent : KernelAgent
         Verify.NotNull(messages);
 
         var chat = new ChatHistory(this._instructions);
-        chat.AddRange(messages.Select(m => this.CreateChatMessage(m)));
+        chat.AddRange(messages.Select(m => CreateChatMessage(m)));
+
+        var chatMessageContent = await this.InvokeAsync(chat, cancellationToken).ConfigureAwait(false);
+
+        return chatMessageContent.Select(m => this.CreateAgentMessage(m)).ToArray();
+    }
+
+    /// <summary>
+    /// Invokes the agent to process the given messages and generate a response.
+    /// </summary>
+    /// <param name="messages">A list of the messages for the agent to process.</param>
+    /// <param name="cancellationToken">An optional <see cref="CancellationToken"/> to cancel the operation.</param>
+    /// <returns>List of messages representing the agent's response.</returns>
+    public async Task<IReadOnlyList<ChatMessageContent>> InvokeAsync(ChatHistory messages, CancellationToken cancellationToken = default)
+    {
+        Verify.NotNull(messages);
 
         var chatCompletionService = this.Kernel.GetRequiredService<IChatCompletionService>();
 
-        var chatMessageContent = await chatCompletionService.GetChatMessageContentsAsync(
-            chat,
+        return await chatCompletionService.GetChatMessageContentsAsync(
+            messages,
             this._promptExecutionSettings,
             this.Kernel,
             cancellationToken).ConfigureAwait(false);
-
-        return chatMessageContent.Select(m => this.CreateAgentMessage(m)).ToArray();
     }
 
     /// <summary>
@@ -54,12 +64,61 @@ public abstract class ChatCompletionAgent : KernelAgent
     /// </summary>
     /// <param name="message">The agent message to be converted.</param>
     /// <returns>A chat message created from the agent message.</returns>
-    protected abstract ChatMessageContent CreateChatMessage(AgentMessage message);
+    private static ChatMessageContent CreateChatMessage(AgentMessage message)
+    {
+        Verify.NotNull(message, nameof(message));
+
+        if (message.Items is not { Count: > 0 })
+        {
+            throw new KernelException("Agent message has no content.");
+        }
+
+        if (message.Role == AuthorRole.User)
+        {
+            return new ChatMessageContent(role: message.Role, items: new(message.Items), innerContent: message);
+        }
+
+        if (message.Role == AuthorRole.Assistant)
+        {
+            if (message.Items.Count != 1)
+            {
+                throw new KernelException("Agent message can't have more than one piece of content for the assistant role.");
+            }
+
+            var content = message.Items.Single();
+
+            if (content is TextContent textContent)
+            {
+                return new ChatMessageContent(
+                    role: message.Role,
+                    content: textContent.Text,
+                    encoding: textContent.Encoding,
+                    metadata: textContent.Metadata,
+                    innerContent: message);
+            }
+
+            throw new KernelException($"Agent message has an unsupported content type '{content.GetType()}' for the assistant role.");
+        }
+
+        return new ChatMessageContent(role: message.Role, items: new(message.Items), innerContent: message);
+    }
 
     /// <summary>
     /// Creates an agent message from a chat message.
     /// </summary>
     /// <param name="message">The chat message to be converted.</param>
     /// <returns>An agent message created from the chat message.</returns>
-    protected abstract AgentMessage CreateAgentMessage(ChatMessageContent message);
+    private AgentMessage CreateAgentMessage(ChatMessageContent message)
+    {
+        Verify.NotNull(message, nameof(message));
+
+        return new AgentMessage(
+            role: message.Role,
+            content: message.Content,
+            innerMessage: message,
+            agent: this);
+    }
+
+    private readonly string _instructions;
+    private readonly PromptExecutionSettings? _promptExecutionSettings;
 }
